Added warning
Recommending the Teensylu not to be used until changes and modifications
are made. This is based on the risk of damage or fire! Either do not use
the board or take proper precautions!
There is no such thing as a "Safe" 3d printer controller. Keep a smoke
detector nearby and have a fire extinguisher handy. Do NOT leave
unattended.